+++ spectrochempy/application.py
@@ -47,8 +47,9 @@ from jinja2 import Template
 from spectrochempy.utils import MetaConfigurable, pathclean, get_pkg_path
 from .matplotlib_preferences import MatplotlibPreferences
 
+
 # set the default style
-plt.style.use(['classic']) #, 'scpy'])    <--- a workaround to solve the problem with
+plt.style.use(['classic'])
 
 # ----------------------------------------------------------------------------------------------------------------------
 # Log levels