+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 * Docker image building
 
+* cantera installation optional
+
+
 
 ## Version 0.2.4